attempts to implement serialization of injector. in progress...
