+++ projects/interactions/private/pybindings/DarkNewsCrossSection.h
@@ -9,366 +9,12 @@
 
 #include "../../public/SIREN/interactions/CrossSection.h"
 #include "../../public/SIREN/interactions/DarkNewsCrossSection.h"
-#include "../../../dataclasses/public/SIREN/dataclasses/Particle.h"
-#include "../../../dataclasses/public/SIREN/dataclasses/InteractionRecord.h"
-#include "../../../dataclasses/public/SIREN/dataclasses/InteractionSignature.h"
-#include "../../../geometry/public/SIREN/geometry/Geometry.h"
-#include "../../../utilities/public/SIREN/utilities/Random.h"
-#include "../../../utilities/public/SIREN/utilities/Pybind11Trampoline.h"
-
-namespace siren {
-namespace interactions {
-// Trampoline class for CrossSection
-class pyCrossSection : public CrossSection,Pybind11Trampoline<CrossSection, pyCrossSection>  {
-public:
-    using CrossSection::CrossSection;
-    pyCrossSection(CrossSection && parent) : CrossSection(std::move(parent)) {}
-    pybind11::object self;
-
-    double TotalCrossSection(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            CrossSection,
-            double,
-            TotalCrossSection,
-            "TotalCrossSection",
-            interaction
-        )
-    }
-
-    double TotalCrossSectionAllFinalStates(siren::dataclasses::InteractionRecord const & record) const override {
-        SELF_OVERRIDE(
-            self,
-            CrossSection,
-            double,
-            TotalCrossSectionAllFinalStates,
-            "TotalCrossSectionAllFinalStates",
-            record
-        )
-    }
-
-    double DifferentialCrossSection(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            CrossSection,
-            double,
-            DifferentialCrossSection,
-            "DifferentialCrossSection",
-            interaction
-        )
-    }
-
-    double InteractionThreshold(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            CrossSection,
-            double,
-            InteractionThreshold,
-            "InteractionThreshold",
-            interaction
-        )
-    }
-
-    void SampleFinalState(dataclasses::CrossSectionDistributionRecord & record, std::shared_ptr<siren::utilities::SIREN_random> random) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            CrossSection,
-            void,
-            SampleFinalState,
-            "SampleFinalState",
-            record,
-            random
-        )
-    }
-
-    std::vector<siren::dataclasses::ParticleType> GetPossibleTargets() const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            CrossSection,
-            std::vector<siren::dataclasses::ParticleType>,
-            GetPossibleTargets,
-            "GetPossibleTargets"
-        )
-    }
-
-    std::vector<siren::dataclasses::ParticleType> GetPossibleTargetsFromPrimary(siren::dataclasses::ParticleType primary_type) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            CrossSection,
-            std::vector<siren::dataclasses::ParticleType>,
-            GetPossibleTargetsFromPrimary,
-            "GetPossibleTargetsFromPrimary",
-            primary_type
-        )
-    }
-
-    std::vector<siren::dataclasses::ParticleType> GetPossiblePrimaries() const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            CrossSection,
-            std::vector<siren::dataclasses::ParticleType>,
-            GetPossiblePrimaries,
-            "GetPossiblePrimaries"
-        )
-    }
-
-    std::vector<siren::dataclasses::InteractionSignature> GetPossibleSignatures() const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            CrossSection,
-            std::vector<siren::dataclasses::InteractionSignature>,
-            GetPossibleSignatures,
-            "GetPossibleSignatures"
-        )
-    }
-
-    std::vector<siren::dataclasses::InteractionSignature> GetPossibleSignaturesFromParents(siren::dataclasses::ParticleType primary_type, siren::dataclasses::ParticleType target_type) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            CrossSection,
-            std::vector<siren::dataclasses::InteractionSignature>,
-            GetPossibleSignaturesFromParents,
-            "GetPossibleSignaturesFromParents",
-            primary_type,
-            target_type
-        )
-    }
-
-    double FinalStateProbability(dataclasses::InteractionRecord const & record) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            CrossSection,
-            double,
-            FinalStateProbability,
-            "FinalStateProbability",
-            record
-        )
-    }
-
-};
-// Trampoline class for DarkNewsCrossSection
-class pyDarkNewsCrossSection : public DarkNewsCrossSection, public Pybind11Trampoline<DarkNewsCrossSection, pyDarkNewsCrossSection> {
-public:
-    using DarkNewsCrossSection::DarkNewsCrossSection;
-    pyDarkNewsCrossSection(DarkNewsCrossSection && parent) : DarkNewsCrossSection(std::move(parent)) {
-        self = pybind11::reinterpret_borrow<pybind11::object>(pybind11::handle(get_object_handle(&parent, pybind11::detail::get_type_info(typeid(DarkNewsCrossSection)))));
-    }
-    pyDarkNewsCrossSection(DarkNewsCrossSection const & parent) : DarkNewsCrossSection(parent) {
-        self = pybind11::reinterpret_borrow<pybind11::object>(pybind11::handle(get_object_handle(&parent, pybind11::detail::get_type_info(typeid(DarkNewsCrossSection)))));
-    }
-    pybind11::object self;
-
-    double TotalCrossSectionAllFinalStates(siren::dataclasses::InteractionRecord const & record) const override {
-        SELF_OVERRIDE(
-            self,
-            CrossSection,
-            double,
-            TotalCrossSectionAllFinalStates,
-            "TotalCrossSectionAllFinalStates",
-            std::cref(record)
-        )
-    }
-
-    double TotalCrossSection(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            double,
-            TotalCrossSection,
-            "TotalCrossSection",
-            std::cref(interaction)
-        )
-    }
-
-    double TotalCrossSection(siren::dataclasses::ParticleType primary, double energy, siren::dataclasses::ParticleType target) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            DarkNewsCrossSection,
-            double,
-            TotalCrossSection,
-            "TotalCrossSection",
-            primary,
-            energy,
-            target
-        )
-    }
-
-    double DifferentialCrossSection(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            double,
-            DifferentialCrossSection,
-            "DifferentialCrossSection",
-            std::cref(interaction)
-        )
-    }
-
-    double DifferentialCrossSection(siren::dataclasses::ParticleType primary, siren::dataclasses::ParticleType target, double energy, double Q2) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            double,
-            DifferentialCrossSection,
-            "DifferentialCrossSection",
-            primary,
-            target,
-            energy,
-            Q2
-        )
-    }
-
-    double InteractionThreshold(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            double,
-            InteractionThreshold,
-            "InteractionThreshold",
-            std::cref(interaction)
-        )
-    }
-
-    double Q2Min(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            double,
-            Q2Min,
-            "Q2Min",
-            std::cref(interaction)
-        )
-    }
-
-    double Q2Max(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            double,
-            Q2Max,
-            "Q2Max",
-            std::cref(interaction)
-        )
-    }
-
-    double TargetMass(dataclasses::ParticleType const & target_type) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            double,
-            TargetMass,
-            "TargetMass",
-            std::cref(target_type)
-        )
-    }
-
-    std::vector<double> SecondaryMasses(std::vector<dataclasses::ParticleType> const & secondary_types) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            std::vector<double>,
-            SecondaryMasses,
-            "SecondaryMasses",
-            std::cref(secondary_types)
-        )
-    }
-
-    std::vector<double> SecondaryHelicities(dataclasses::InteractionRecord const & record) const override{
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            std::vector<double>,
-            SecondaryHelicities,
-            "SecondaryHelicities",
-            std::cref(record)
-        )
-    }
-
-    void SampleFinalState(dataclasses::CrossSectionDistributionRecord & record, std::shared_ptr<siren::utilities::SIREN_random> random) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            void,
-            SampleFinalState,
-            "SampleFinalState",
-            std::ref(record),
-            random
-        )
-    }
-
-    std::vector<siren::dataclasses::ParticleType> GetPossibleTargets() const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            DarkNewsCrossSection,
-            std::vector<siren::dataclasses::ParticleType>,
-            GetPossibleTargets,
-            "GetPossibleTargets"
-        )
-    }
-
-    std::vector<siren::dataclasses::ParticleType> GetPossibleTargetsFromPrimary(siren::dataclasses::ParticleType primary_type) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            DarkNewsCrossSection,
-            std::vector<siren::dataclasses::ParticleType>,
-            GetPossibleTargetsFromPrimary,
-            "GetPossibleTargetsFromPrimary",
-            primary_type
-        )
-    }
-
-    std::vector<siren::dataclasses::ParticleType> GetPossiblePrimaries() const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            DarkNewsCrossSection,
-            std::vector<siren::dataclasses::ParticleType>,
-            GetPossiblePrimaries,
-            "GetPossiblePrimaries"
-        )
-    }
-
-    std::vector<siren::dataclasses::InteractionSignature> GetPossibleSignatures() const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            DarkNewsCrossSection,
-            std::vector<siren::dataclasses::InteractionSignature>,
-            GetPossibleSignatures,
-            "GetPossibleSignatures"
-        )
-    }
-
-    std::vector<siren::dataclasses::InteractionSignature> GetPossibleSignaturesFromParents(siren::dataclasses::ParticleType primary_type, siren::dataclasses::ParticleType target_type) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            DarkNewsCrossSection,
-            std::vector<siren::dataclasses::InteractionSignature>,
-            GetPossibleSignaturesFromParents,
-            "GetPossibleSignaturesFromParents",
-            primary_type,
-            target_type
-        )
-    }
-
-    double FinalStateProbability(dataclasses::InteractionRecord const & record) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsCrossSection,
-            double,
-            FinalStateProbability,
-            "FinalStateProbability",
-            std::cref(record)
-        )
-    }
-
-};
-} // end interactions namespace
-} // end LI namespace
 
 void register_DarkNewsCrossSection(pybind11::module_ & m) {
     using namespace pybind11;
     using namespace siren::interactions;
 
-    class_<DarkNewsCrossSection, std::shared_ptr<DarkNewsCrossSection>, CrossSection, siren::interactions::pyDarkNewsCrossSection> DarkNewsCrossSection(m, "DarkNewsCrossSection");
+    class_<DarkNewsCrossSection, std::shared_ptr<DarkNewsCrossSection>, siren::interactions::pyDarkNewsCrossSection, CrossSection> DarkNewsCrossSection(m, "DarkNewsCrossSection");
 
     DarkNewsCrossSection
         .def(init<>())
